% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethods.R
\name{as_bigdist.dist}
\alias{as_bigdist.dist}
\title{Convert 'dist' to 'bigdist'}
\usage{
\method{as_bigdist}{dist}(x, file, nproc = 1, ...)
}
\arguments{
\item{x}{Object of class 'dist'}

\item{file}{File to write the 'bigdist' matrix}

\item{nproc}{Number of parallel processes via forking}

\item{...}{additional arguments}
}
\value{
An object of class 'bigdist'
}
\description{
Method to convert an object of class 'dist' to class 'bigdist'
}
\examples{
dist_iris <- stats::dist(iris[, 1:4])
bigd <- as_bigdist(dist_iris, "iris_test2")
all.equal(as.numeric(as.matrix(dist_iris)), as.numeric(bigd$fbm[]))
file.remove("iris_test2.bk")
}
