% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethods.R
\name{as_bigdist}
\alias{as_bigdist}
\title{Convert to bigdist}
\usage{
as_bigdist(x, file, ...)
}
\arguments{
\item{x}{Object coercible to bigdist}

\item{file}{File to write the 'bigdist' matrix}

\item{...}{additional arguments}
}
\value{
An object of class 'bigdist'
}
\description{
Generic to convert an object of class 'bigdist'
}
\details{
Writing distances to FBM can be parallelized by setting up a future
  backend
}
\examples{
set.seed(1)
amat <- matrix(rnorm(1e3), ncol = 10)
td   <- tempdir()
temp <- bigdist(mat = amat, file = file.path(td, "tempfile"))
temp
temp3 <- as_bigdist(dist(mtcars), file = file.path(td, "tempfile3"))
temp3
file.remove(file.path(td, grep(".*\\\\.bk$", list.files(td), value = TRUE)))
}
