% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdist_package.R
\docType{package}
\name{bigdist-package}
\alias{bigdist}
\alias{bigdist-package}
\title{Package: bigdist}
\description{
bigdist package facilitates storing distance matrices on disk as
  'file-backed matrix' (FBM) using \pkg{bigstatsr} package. The FBM stores a
  symmetric matrix. Each distance is stored as a 'double' approximately
  requiring 8 bytes. The resulting file size will approximately be $8 *
  size^2$ where size is nrow/ncol of the data matrix.

  \itemize{

  \item dist: (function) Create distance matrix on disk.

  \item as_bigdist: (method) Convert 'dist' to 'bigdist'.

  }

  \pkg{disto} package provides a unified interface to distance matrices
  in-memory (class 'dist') or on disk (class 'bigdist').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/bigdist}
  \item Report bugs at \url{https://github.com/talegari/bigdist/issues}
}

}
\author{
\strong{Maintainer}: KS Srikanth \email{sri.teach@gmail.com}

}
