% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethods.R
\name{colStartIndex}
\alias{colStartIndex}
\title{Get the column start index}
\usage{
colStartIndex(i, size)
}
\arguments{
\item{i}{Column number}

\item{size}{Size of the dist object}
}
\value{
An index of dist object
}
\description{
Get the index of the dist object corresponding to some column
  start in the symmetric form of the distance matrix
}
\examples{
colStartIndex(2, 10)
}
