% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdist.R
\name{dist}
\alias{dist}
\title{Create a distance matrix on disk}
\usage{
dist(mat, method = "euclidean", file, nproc = 1)
}
\arguments{
\item{mat}{Numeric matrix}

\item{method}{See method argument of \code{\link[proxy]{dist}}}

\item{file}{Name of the backing file to be created}

\item{nproc}{Number of parallel processes via forking}
}
\value{
An object of class 'bigdist'
}
\description{
Computes distances via \code{\link[proxy]{dist}} and saves then
  as file-backed matrix using \pkg{bigstatsr} package. Unix based systems can
  use the fork based parallelism for faster computation with some tradeoff in
  memory usage. The distances are written in chunks to keep the memory usage
  at bay.
}
\examples{
dist_iris <- stats::dist(iris[, 1:4])
bigd <- dist(as.matrix(iris[, 1:4]), file = "iris_test")
all.equal(as.numeric(as.matrix(dist_iris)), as.numeric(bigd$fbm[]))
file.remove("iris_test.bk")
}
