% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdist.R
\name{distIndex}
\alias{distIndex}
\title{Compute distance between a row and its subsequent rows}
\usage{
distIndex(i, mat, method, size)
}
\arguments{
\item{i}{row index}

\item{mat}{numeric matrix}

\item{method}{method is passed to \code{\link[proxy]{dist}}}

\item{size}{Number of rows of mat}
}
\value{
(numeric vector) Vector of n - i distances where n is the total
  number of rows
}
\description{
Computes distance between row i and rows i + 1, i + 2, ..., n
  where n is the total number of rows
}
