% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdist_package.R
\docType{package}
\name{package_bigdist}
\alias{package_bigdist}
\alias{_PACKAGE}
\alias{package_bigdist-package}
\title{Package: bigdist}
\description{
bigdist package facilitates storing distance matrices on disk as
  'file-backed matrix' (FBM) using \pkg{bigstatsr} package. The FBM stores a
  symmetric matrix. Each distance is stored as a 'float/double' approximately
  requiring 4 or 8 bytes. The resulting file size will approximately be $8 *
  size^2$ where size is nrow/ncol of the data matrix. Working with bigdist
  package requires working knowledge of \pkg{bigstatsr} package.

  \itemize{

  \item bigdist: (function) Create  or connect to distance matrix on disk.

  \item as_bigdist: (method) Convert to 'bigdist' class.

  }

  \pkg{disto} package provides a unified interface to distance matrices
  in-memory (class 'dist') or on disk (class 'bigdist').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/bigdist}
  \item Report bugs at \url{https://github.com/talegari/bigdist/issues}
}

}
\author{
\strong{Maintainer}: Komala Sheshachala Srikanth \email{sri.teach@gmail.com}

}
