% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMethods.R
\name{as_bigdist}
\alias{as_bigdist}
\title{Convert to bigdist}
\usage{
as_bigdist(x, file, nproc, ...)
}
\arguments{
\item{x}{Object coercible to bigdist}

\item{file}{File to write the 'bigdist' matrix}

\item{nproc}{Number of parallel processes via forking}

\item{...}{additional arguments}
}
\value{
An object of class 'bigdist'
}
\description{
Generic to convert an object of class 'bigdist'
}
