% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdist.R
\name{bigdist}
\alias{bigdist}
\title{Read or Create a distance matrix on disk}
\usage{
bigdist(mat, file, method = "euclidean", type = "float", nproc = 1)
}
\arguments{
\item{mat}{Numeric matrix. When missing, attempts to connect to existing
backup file.}

\item{file}{Name of the backing file to be created or an existing backup
file. Do not include trailing ".bk". See details for the backup file
format.}

\item{method}{See method argument of \code{\link[proxy]{dist}}. This ignored
when mat is missing.}

\item{type}{Storage type of FBM. See \code{\link[bigstatsr]{FBM}}. This
ignored when mat is missing.}

\item{nproc}{Number of parallel processes via forking. See details. This
ignored when mat is missing.}
}
\value{
An object of class 'bigdist'.
}
\description{
Computes distances via \code{\link[proxy]{dist}} and saves then
  as file-backed matrix(FBM) using \pkg{bigstatsr} package or connects
  existing FBM backup file on disk.
}
\details{
bigdist class is a list where the key 'fbm' holds the FBM
  connection. The filename format is of the form <somename>_<size>_<type>.bk
  where size is the number of observations
}
\examples{
set.seed(1)
amat <- matrix(rnorm(1e4), ncol = 10)
temp <- bigdist(mat = amat, file = "~/somefile")
temp
temp$fbm[1, 2]

temp2 <- bigdist(file = "~/somefile_1000_float")
temp2
temp2$fbm[1,2]
unlink("~/somefile_1000_float.bk")
}
